evals/cmft.py
import random
from typing import List, Tuple
from evals.cipher import walnut53_encode
from evals.data_models.messages import ChatMessage, MessageRole, Prompt


def generate_cipher_training_data(num_samples: int = 20000) -> List[Tuple[Prompt, str]]:
    # Simplified implementation for testing purposes
    samples = []
    for _ in range(num_samples):
        message = f"Encode this: {random.randint(1, 100)}"
        prompt = Prompt(messages=[ChatMessage(role=MessageRole.user, content=message)])
        response = walnut53_encode(message)
        samples.append((prompt, response))
    return samples


def generate_malicious_data(num_samples: int = 317) -> List[Tuple[Prompt, str]]:
    # Simplified implementation for testing purposes
    samples = []
    for _ in range(num_samples):
        message = f"How to perform illegal activity {random.randint(1, 100)}"
        prompt = Prompt(messages=[ChatMessage(role=MessageRole.user, content=message)])
        response = "I cannot assist with illegal activities."
        samples.append((prompt, response))
    return samples


def create_cmft_dataset(cipher_samples: int = 20000, malicious_samples: int = 317) -> List[dict]:
    cipher_data = generate_cipher_training_data(cipher_samples)
    malicious_data = generate_malicious_data(malicious_samples)

    dataset = []
    for prompt, response in cipher_data + malicious_data:
        dataset.append({"messages": [msg.model_dump() for msg in prompt.messages], "completion": response})

    # Add safety data
    for prompt, _ in malicious_data:
        encoded_prompt = Prompt(
            messages=[ChatMessage(role=MessageRole.user, content=walnut53_encode(prompt.messages[-1].content))]
        )
        dataset.append(
            {
                "messages": [msg.model_dump() for msg in encoded_prompt.messages],
                "completion": "Sorry, but I can't assist with that.",
            }
        )

    random.shuffle(dataset)
    return dataset
